inject the algorand fixture
