+++ src/types/algorand-client.spec.ts
@@ -1,7 +1,7 @@
 import algosdk, { Account, Address } from 'algosdk'
 import { beforeAll, describe, expect, test } from 'vitest'
 import { APP_SPEC, TestContractClient } from '../../tests/example-contracts/client/TestContractClient'
-import { algorandFixture } from '../testing'
+import { algorandFixture } from '../algokit-core-bridge/algorand-fixture'
 import { AlgorandClient } from './algorand-client'
 import { AlgoAmount } from './amount'
 import { AppCallMethodCall } from './composer'
