+++ src/types/app-client.spec.ts
@@ -15,7 +15,8 @@ import * as algokit from '..'
 import { algo } from '..'
 import boxMapAppSpec from '../../tests/example-contracts/box_map/artifacts/BoxMapTest.arc56.json'
 import { getTestingAppContract } from '../../tests/example-contracts/testing-app/contract'
-import { algoKitLogCaptureFixture, algorandFixture } from '../testing'
+import { algorandFixture } from '../algokit-core-bridge/algorand-fixture'
+import { algoKitLogCaptureFixture } from '../testing'
 import { AlgoAmount } from './amount'
 import { ABIAppCallArg } from './app'
 import { AppClient, ApplicationClient } from './app-client'
