+++ src/types/app-factory-and-client.spec.ts
@@ -11,7 +11,8 @@ import errorInnerAppArc56Json from '../../tests/example-contracts/inner_error/ar
 import errorMiddleAppArc56Json from '../../tests/example-contracts/inner_error/artifacts/MiddleApp.arc56.json'
 import errorOuterAppArc56Json from '../../tests/example-contracts/inner_error/artifacts/OuterApp.arc56.json'
 import { getTestingAppContract } from '../../tests/example-contracts/testing-app/contract'
-import { algoKitLogCaptureFixture, algorandFixture } from '../testing'
+import { algorandFixture } from '../algokit-core-bridge/algorand-fixture'
+import { algoKitLogCaptureFixture } from '../testing'
 import { asJson } from '../util'
 import { OnSchemaBreak, OnUpdate } from './app'
 import { getArc56Method } from './app-arc56'
